+++ my/src/App.js
@@ -1,73 +1,20 @@
 import './App.css';
-import {BrowserRouter, Routes, Route} from 'react-router-dom'
-import ReturnPage from './ReturnPage'
-import NotFound from './NotFound'
+import {BrowserRouter, Routes, Route} from 'react-router-dom';
+import ReturnPage from './ReturnPage';
+import NotFound from './NotFound';
+import Main from './Main';
 
-const data = [
-  {
-    albumId: 1,
-    id: 1,
-    title: "accusamus beatae ad facilis cum similique qui sunt",
-    url: "https://via.placeholder.com/600/92c952",
-    thumbnailUrl: "https://via.placeholder.com/150/92c952",
-  },
-  {
-    albumId: 1,
-    id: 2,
-    title: "reprehenderit est deserunt velit ipsam",
-    url: "https://via.placeholder.com/600/771796",
-    thumbnailUrl: "https://via.placeholder.com/150/771796",
-  },
-  {
-    albumId: 1,
-    id: 3,
-    title: "officia porro iure quia iusto qui ipsa ut modi",
-    url: "https://via.placeholder.com/600/24f355",
-    thumbnailUrl: "https://via.placeholder.com/150/24f355",
-  },
-  {
-    albumId: 1,
-    id: 4,
-    title: "culpa odio esse rerum omnis laboriosam voluptate repudiandae",
-    url: "https://via.placeholder.com/600/d32776",
-    thumbnailUrl: "https://via.placeholder.com/150/d32776",
-  },
-  {
-    albumId: 1,
-    id: 5,
-    title: "natus nisi omnis corporis facere molestiae rerum in",
-    url: "https://via.placeholder.com/600/f66b97",
-    thumbnailUrl: "https://via.placeholder.com/150/f66b97",
-  },
-  {
-    albumId: 1,
-    id: 6,
-    title: "accusamus ea aliquid et amet sequi nemo",
-    url: "https://via.placeholder.com/600/56a8c2",
-    thumbnailUrl: "https://via.placeholder.com/150/56a8c2",
-  },
-];
 
 const App = () => {
   return (
-    
-    <div className="back">
-      <BrowserRouter>
-      {data.map((v) => (
-        <div className="component" key={v.title}>
-          <Routes>
-          
-          <img src={v.thumbnailUrl} alt="1" />
-          <Route path='v.thumbnailUrl' element={<ReturnPage />}></Route>
-          <Route path='*' element={<NotFound />}></Route>
-          <div>
-            {v.title} / {v.id}
-          </div>
-          </Routes>
-        </div>
-      ))}
-      </BrowserRouter>
-    </div>
+    <BrowserRouter>
+      <Main />
+      <Routes>
+        <Route path='/' element={<Main />}></Route>
+        <Route path='/main.thumbnailUrl' element={<ReturnPage />}></Route>
+        <Route path='/*' element={<NotFound />}></Route>
+      </Routes>
+    </BrowserRouter>
     
   );
 };
